Update inizio gennaio 2025
aggiornamento e/o commit dei seguenti progetti:

OSINode
centralina delay
EP181 (poi abbiamo usato il codice di Emiliano)
EP343 lettore tensioni per Andrea (da perfezionare, forse)
EP417 prototipo
LED_STRISCIA test con libreria NeoPixel
OSINode Carnemolla

+++ 4FoxNode_ottobre_2024/include/sht3x.h
@@ -1,1 +1,1 @@
-unsigned short int *sht3x(int);
+float *sht3x(int);
